Change deploy command from a subcommand to a main command
