--- plugins/kn-services/cmd/root.go
@@ -1,79 +1,0 @@
-/*
-Copyright © 2019 NAME HERE <EMAIL ADDRESS>
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package cmd
-
-import (
-  "fmt"
-  "os"
-  "github.com/spf13/cobra"
-
-  homedir "github.com/mitchellh/go-homedir"
-  "github.com/spf13/viper"
-
-)
-
-
-var cfgFile string
-var kubeconfig string
-
-
-var rootCmd = &cobra.Command{
-  Use:   "kn-service",
-  Short: "Build or deploy knative application",
-  Long: `Build or deploy knative application`,
-}
-
-func Execute() {
-  if err := rootCmd.Execute(); err != nil {
-    fmt.Println(err)
-    os.Exit(1)
-  }
-}
-
-func init() {
-  cobra.OnInitialize(initConfig)
-
-  rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.app.yaml)")
-  rootCmd.PersistentFlags().StringP("kubeconfig", "", "", "kube config file (default is KUBECONFIG from ENV property)")
-}
-
-
-// initConfig reads in config file and ENV variables if set.
-func initConfig() {
-  if cfgFile != "" {
-    // Use config file from the flag.
-    viper.SetConfigFile(cfgFile)
-  } else {
-    // Find home directory.
-    home, err := homedir.Dir()
-    if err != nil {
-      fmt.Println(err)
-      os.Exit(1)
-    }
-
-    // Search config in home directory with name ".app" (without extension).
-    viper.AddConfigPath(home)
-    viper.SetConfigName(".app")
-  }
-
-  viper.AutomaticEnv() // read in environment variables that match
-
-  // If a config file is found, read it in.
-  if err := viper.ReadInConfig(); err == nil {
-    fmt.Println("Using config file:", viper.ConfigFileUsed())
-  }
-}
-
